Move all visualization/pyglet functionalities to a dedicated file
