+++ pywavefront/parser.py
@@ -33,7 +33,6 @@
 # ----------------------------------------------------------------------------
 
 import os
-import pyglet
 import logging
 
 class Parser(object):
